requirements.txt
PySide6>=6.9.1
PySide6_Addons>=6.9.1
PySide6_Essentials>=6.9.1
python3-xlib>=0.15
shiboken6>=6.9.1